+++ component/component.go
@@ -25,9 +25,9 @@ const (
 	NSAccept = `jabber:component:accept`
 )
 
-// NewClientSession initiates an XMPP session on the given io.ReadWriter using
-// the component protocol.
-func NewClientSession(ctx context.Context, addr jid.JID, secret []byte, rw io.ReadWriter, received bool) (*xmpp.Session, error) {
+// NewSession initiates an XMPP session on the given io.ReadWriter using the
+// component protocol.
+func NewSession(ctx context.Context, addr jid.JID, secret []byte, rw io.ReadWriter, received bool) (*xmpp.Session, error) {
 	addr = addr.Domain()
 	if received {
 		return xmpp.ReceiveSession(ctx, rw, Negotiator(addr, secret, received))