+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking
 
+- component: `NewClientSession` has been renamed `NewSession`
 - mux: fixed the signature of the `IQFunc` option to differentiate it from the
   normal `IQ` option
 - stream: the `SeeOtherHost` function signature has changed now that payload is